:gear: Change GitHub Action to Push to GHCR
Unfortunately the Upstream Repo has been having some build & push
failures over the last year and they haven't actively made any changes
to fix that.

As our fork is updated and in sync with latest, we could essentially
replace the action and push to GHCR to be able to deploy the latest
image.

**TL;DR**
This will replace the upstream Github Action to push to the
`fairfaxmedia/confidant` GHCR instead of Docker Hub